// src/app/(pages)/offering/types.ts
export type OfferingDetail = {
  productId: number
  productName: string
  owner?: string
  thumbnailImg?: string
  presentImg?: string
  detailImg?: string
  projectName?: string
  issueAmount?: number
  tokenStandard?: string
  exchangeListing?: string
  tokenQuantity?: number
  tokenName?: string
  progressRate?: number
  offeringStartDate?: string
  offeringEndDate?: string
  offeringPrice?: number
  offeringAmount?: number
  offeringTokens?: number
  isFavorite?: boolean
}